some fixes and improvements

+++ app/src/parser/outputs/secp_output_owners.rs
@@ -96,7 +96,7 @@ impl<'a> DisplayableItem for SECPOutputOwners<'a> {
         use bolos::{pic_str, PIC};
         use lexical_core::{write as itoa, Number};
 
-        let mut buffer = [0; usize::FORMATTED_SIZE + 2];
+        let mut buffer = [0; u64::FORMATTED_SIZE + 2];
         let addr_item_n = self.num_items() - self.addresses.len();
 
         match item_n as usize {
